+++ src/drawing/DrawingPlaneSettings.java
@@ -2,9 +2,9 @@ package drawing;
 
 public class DrawingPlaneSettings {
 
-    public int planeX;
-    public int planeY;
-    public int tileSize;
+    public float planeX;
+    public float planeY;
+    public float tileSize;
 
     public DrawingPlaneSettings() {}
